# docs/swagger.yaml
definitions:
...
  controllers.categoryInput:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  controllers.orderInput:
    properties:
      courier:
        type: string
      payment:
        type: string
      user_id:
        type: integer
    type: object
  controllers.orderSendInput:
    properties:
      resi:
        type: string
    type: object
...
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Order:
    properties:
      complete_at:
        type: string
      courier:
        type: string
      created_at:
        type: string
      id:
        type: integer
      paid_at:
        type: string
      payment:
        type: string
      resi:
        type: string
      sent_at:
        type: string
      status:
        type: string
      total:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      photo_path:
        type: string
      price:
        type: integer
      shop_id:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
      weight:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /categories:
    get:
      description: Get a list of Category
      parameters:
      - description: name search by keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all Category
      tags:
      - Public
    post:
      description: Creating new Category
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create new Category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.categoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Create new Category
      tags:
      - Admin
  /categories/{id}:
    delete:
      description: Delete category by id
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete category
      tags:
      - Admin
    put:
      description: Update Category by id
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.categoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Update Category
      tags:
      - Admin
  /categories/{id}/products:
    get:
      description: Get all Products by CategoryId
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get Products
      tags:
      - Public
...
  /orders:
    get:
      description: Get a list of Order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
      summary: Get all Order
      tags:
      - Buyer
      - Admin
    post:
      description: Creating new Order
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create new Order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.orderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Create new Order
      tags:
      - Buyer
  /orders/{id}:
    delete:
      description: Delete order by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete order
      tags:
      - Admin
    get:
      description: Get Order by Id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Get Order detail
      tags:
      - Buyer
      - Admin
    put:
      description: Update Order by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.orderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Update Order
      tags:
      - Buyer
      - Admin
  /orders/{id}/complete:
    post:
      description: Complete Order by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Complete Order
      tags:
      - Buyer
  /orders/{id}/pay:
    post:
      description: Pay Order by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Pay Order
      tags:
      - Buyer
  /orders/{id}/send:
    post:
      description: Send Order by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.orderSendInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Send Order
      tags:
      - Seller
      - Admin
...
swagger: "2.0"